AB2D-4424: Move ZIPFORMAT out of JobService (in order to break dependencies on JobService).
