+++ common/src/main/java/gov/cms/ab2d/common/util/Constants.java
@@ -79,4 +79,6 @@ public final class Constants {
     // wanted to include it in the swagger documentation and for the swagger annotation, the value has to be
     // constant at compile time so I put it here.
     public static final String SINCE_EARLIEST_DATE = "2020-02-13T00:00:00.000-05:00";
+
+    public static final String ZIPFORMAT = "application/zip";
 }
